rename folder/js/content-grid to components, move content-grid-js to /js/ folder
