+++ src/js/components/loader-spinner.js
