+++ src/js/components/update-genres-info.js
