+++ src/js/components/update-rating.js
