+++ src/js/components/update-year-info.js
