+++ src/js/main.js
@@ -1,2 +1,2 @@
 import "./header/header.main";
-import "./content-grid/content-grid.main";
+import "./content-grid.main";
